refactor(CardScoreResult): 클래스명 수정
